+++ lib/example/example.dart
@@ -9,7 +9,7 @@ import 'package:sample_exception/sample_exception.dart';
 
 /// example
 void main() async {
-  final exception = SampleException(
+  final exception = SampleException<Exception>(
     SampleErrorCode.unknown,
     error: Exception("Sample error"),
     message: 'Test sample error',
@@ -17,10 +17,18 @@ void main() async {
     time: DateTime.now(),
   );
 
-  final baseon = SampleException('hi', error: exception);
-  final baseon2 = NetworkException(error: exception);
-  print(baseon);
-  print(baseon2);
+  final baseon = SampleException(SampleErrorCode.unknown, error: exception);
+  final baseon2 = NetworkException(error: baseon);
+  final baseon3 = BlocException(error: baseon2);
+  final rs1 = baseon3.containsError('hi');
+  final rs2 = baseon3.whereError<NetworkException>('network-');
+  final rs3 = baseon3.baseError();
+  final rs4 = baseon3.checkBaseError('hi');
+  final rs5 = baseon3.checkBaseError(SampleErrorCode.unknown);
+  final rs6 = baseon3.whereError('test');
+  final rs7 =
+      baseon3.whereError<SampleException<Exception>>(SampleErrorCode.unknown);
+  print(baseon3);
   // print('Print: ');
   // exception.print();
   // print('Throw: ');
@@ -41,4 +49,20 @@ class NetworkException extends SampleException {
   @override
   // TODO: implement errorCode
   String get errorCode => 'network-${super.errorCode}';
+}
+
+class BlocException extends SampleException {
+  BlocException({
+    String errorCode = '',
+    dynamic error,
+    StackTrace? stackTrace,
+  }) : super(errorCode,
+            error: error,
+            stackTrace: stackTrace,
+            time: DateTime.now(),
+            message: 'Error in bloc:');
+
+  @override
+  // TODO: implement errorCode
+  String get errorCode => 'bloc-${super.errorCode}';
 }