+++ lib/sample_exception.dart
@@ -1,3 +1,4 @@
 library sample_exception;
 
 export 'src/sample_exception.dart';
+export 'src/sample_error.dart';