+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.0.1
+
+- Remove tree constructor in SampleException.errorCode
+- Add tree constructor as field to SampleException